feat: Start scenario A
Scenario A: A form is presented to the user and some form fields trigger
adding additional form fields via htmx, e.g. when entering a zip code, a
htmx request is made to get a selection of network operators that is
then inserted into the form.
In the end, all form data is submitted, and the processing view can
access all the information (distributing it over different models if
necessary)